+++ R/load.R
@@ -149,11 +149,9 @@ load_all <- function(path = ".",
   }
 
   if (isTRUE(compile)) {
-    rlang::check_installed("pkgbuild", reason = "to compile packages with a `src/` directory.")
     pkgbuild::clean_dll(path)
     pkgbuild::compile_dll(path, quiet = quiet)
   } else if (identical(compile, NA)) {
-    rlang::check_installed("pkgbuild", reason = "to compile packages with a `src/` directory.")
     pkgbuild::compile_dll(path, quiet = quiet)
   } else if (identical(compile, FALSE)) {
     # don't compile