+++ NEWS.md
@@ -1,3 +1,5 @@
+# pkgload (development version)
+
 # pkgload 1.3.2
 
 * Fixes for CRAN checks.