+++ NEWS.md
@@ -1,9 +1,12 @@
 # pkgload (development version)
 
-* `load_all()` no longer standardises version number in namespace metadata 
+* pkgload now depends unconditionally on pkgbuild (#249).
+
+* `load_all()` no longer standardises version number in namespace metadata
   (#231).
-  
+
 * New `pkg_version_raw()` to get raw package version as a string.
+
 
 # pkgload 1.3.2